filters: Convert existing filters to use OnBuildFilter
