+++ pkg/filters/ip.go
@@ -1,0 +1,82 @@
+// Copyright 2019-2020 Authors of Hubble
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package filters
+
+import (
+	"context"
+	"fmt"
+	"net"
+
+	pb "github.com/cilium/hubble/api/v1/flow"
+	v1 "github.com/cilium/hubble/pkg/api/v1"
+)
+
+func sourceIP(ev *v1.Event) string {
+	return ev.GetFlow().GetIP().GetSource()
+}
+
+func destinationIP(ev *v1.Event) string {
+	return ev.GetFlow().GetIP().GetDestination()
+}
+
+func filterByIPs(ips []string, getIP func(*v1.Event) string) (FilterFunc, error) {
+	for _, ip := range ips {
+		if net.ParseIP(ip) == nil {
+			return nil, fmt.Errorf("invalid IP address in filter: %q", ip)
+		}
+	}
+
+	return func(ev *v1.Event) bool {
+		eventIP := getIP(ev)
+		if eventIP == "" {
+			return false
+		}
+
+		for _, ip := range ips {
+			if ip == eventIP {
+				return true
+			}
+		}
+
+		return false
+	}, nil
+}
+
+// IPFilter implements IP addressing filtering for the source and destination
+// address
+type IPFilter struct{}
+
+// OnBuildFilter builds an IP address filter
+func (f *IPFilter) OnBuildFilter(ctx context.Context, ff *pb.FlowFilter) ([]FilterFunc, error) {
+	var fs []FilterFunc
+
+	if ff.GetSourceIp() != nil {
+		ipf, err := filterByIPs(ff.GetSourceIp(), sourceIP)
+		if err != nil {
+			return nil, err
+		}
+		fs = append(fs, ipf)
+	}
+
+	if ff.GetDestinationIp() != nil {
+		ipf, err := filterByIPs(ff.GetDestinationIp(), destinationIP)
+		if err != nil {
+			return nil, err
+		}
+		fs = append(fs, ipf)
+	}
+
+	return fs, nil
+}
