+++ pkg/filters/verdict.go
@@ -1,0 +1,52 @@
+// Copyright 2019-2020 Authors of Hubble
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package filters
+
+import (
+	"context"
+
+	pb "github.com/cilium/hubble/api/v1/flow"
+	v1 "github.com/cilium/hubble/pkg/api/v1"
+)
+
+func filterByVerdicts(vs []pb.Verdict) FilterFunc {
+	return func(ev *v1.Event) bool {
+		flow := ev.GetFlow()
+		if flow == nil {
+			return false
+		}
+		for _, verdict := range vs {
+			if verdict == flow.GetVerdict() {
+				return true
+			}
+		}
+
+		return false
+	}
+}
+
+// VerdictFilter implements filtering based on forwarding verdict
+type VerdictFilter struct{}
+
+// OnBuildFilter builds a forwarding verdict filter
+func (v *VerdictFilter) OnBuildFilter(ctx context.Context, ff *pb.FlowFilter) ([]FilterFunc, error) {
+	var fs []FilterFunc
+
+	if ff.GetVerdict() != nil {
+		fs = append(fs, filterByVerdicts(ff.GetVerdict()))
+	}
+
+	return fs, nil
+}
