+++ pkg/server/local_observer.go
@@ -220,7 +220,17 @@ func (s *LocalObserverServer) GetFlows(
 	req *observer.GetFlowsRequest,
 	server observer.Observer_GetFlowsServer,
 ) (err error) {
-	return getFlows(req, server, s)
+	filterList := append(filters.DefaultFilters, s.opts.OnBuildFilter...)
+	whitelist, err := filters.BuildFilterList(server.Context(), req.Whitelist, filterList)
+	if err != nil {
+		return err
+	}
+	blacklist, err := filters.BuildFilterList(server.Context(), req.Blacklist, filterList)
+	if err != nil {
+		return err
+	}
+
+	return getFlows(ctx, req, server, s, whitelist, blacklist)
 }
 
 func getFlows(
